refactor: pages에서 관리하던 SignUp을 views에서 관리합니다
